Replace original script
